# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ClassLevel = typeof ClassLevel[keyof typeof ClassLevel];


export const ClassLevel = {
  '6-8': '6-8',
  '9-10': '9-10',
  '11-12': '11-12',
} as const;

export type StudentStatus = typeof StudentStatus[keyof typeof StudentStatus];


export const StudentStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
  terminated: 'terminated',
} as const;

export type FeeStatus = typeof FeeStatus[keyof typeof FeeStatus];


export const FeeStatus = {
  paid: 'paid',
  partial: 'partial',
  pending: 'pending',
} as const;

export type PaymentMode = typeof PaymentMode[keyof typeof PaymentMode];


export const PaymentMode = {
  cash: 'cash',
  upi: 'upi',
  bank_transfer: 'bank_transfer',
  debit_card: 'debit_card',
  credit_card: 'credit_card',
  other: 'other',
} as const;

export interface Student {
  id: number;
  studentId: string;
  admissionNumber: string;
  fullName: string;
  fatherName: string;
  mobile: string;
  /** @nullable */
  email: string | null;
  pinCode: string;
  schoolName: string;
  className: ClassLevel;
  admissionDate: string;
  registrationDate: string;
  status: StudentStatus;
  /** @nullable */
  profilePhotoUrl: string | null;
  /** @nullable */
  terminationDate: string | null;
  /** @nullable */
  terminationReason: string | null;
  createdAt: string;
}

export interface StudentRegistrationInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 1 */
  fatherName: string;
  /**
     * @minLength 10
     * @maxLength 10
     */
  mobile: string;
  email?: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  pinCode: string;
  /** @minLength 1 */
  schoolName: string;
  className: ClassLevel;
  /** @minLength 8 */
  password: string;
}

export interface StudentLoginInput {
  mobile: string;
  password: string;
}

export interface StudentCreateInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 1 */
  fatherName: string;
  /**
     * @minLength 10
     * @maxLength 10
     */
  mobile: string;
  email?: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  pinCode: string;
  /** @minLength 1 */
  schoolName: string;
  className: ClassLevel;
  /** @minLength 8 */
  password: string;
  admissionDate: string;
}

export interface StudentUpdate {
  fullName?: string;
  fatherName?: string;
  mobile?: string;
  email?: string;
  pinCode?: string;
  schoolName?: string;
  className?: ClassLevel;
  admissionDate?: string;
  status?: StudentStatus;
}

export interface ProfilePhotoUpdate {
  profilePhotoUrl: string;
}

export interface PasswordChangeInput {
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface TerminationInput {
  reason: string;
}

export interface Admin {
  id: number;
  username: string;
  mustChangePassword: boolean;
  /** @nullable */
  lastLoginAt: string | null;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

/**
 * @nullable
 */
export type FeePaymentMode = typeof FeePaymentMode[keyof typeof FeePaymentMode] | null;


export const FeePaymentMode = {
  cash: 'cash',
  upi: 'upi',
  bank_transfer: 'bank_transfer',
  debit_card: 'debit_card',
  credit_card: 'credit_card',
  other: 'other',
} as const;

export interface Fee {
  id: number;
  studentId: number;
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  year: number;
  feeAmount: number;
  discount: number;
  lateFee: number;
  amountPaid: number;
  status: FeeStatus;
  /** @nullable */
  paymentDate: string | null;
  /** @nullable */
  receiptNumber: string | null;
  /** @nullable */
  paymentMode: FeePaymentMode;
  /** @nullable */
  remarks: string | null;
}

export interface FeeUpdate {
  feeAmount?: number;
  discount?: number;
  lateFee?: number;
  amountPaid?: number;
  status?: FeeStatus;
  paymentDate?: string;
  paymentMode?: PaymentMode;
  remarks?: string;
}

export interface Note {
  id: number;
  title: string;
  subject: string;
  className: ClassLevel;
  fileUrl: string;
  uploadDate: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  subject: string;
  className: ClassLevel;
  /** @minLength 1 */
  fileUrl: string;
}

export interface Paper {
  id: number;
  title: string;
  board: string;
  subject: string;
  year: number;
  className: ClassLevel;
  fileUrl: string;
  uploadDate: string;
}

export interface PaperInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  board: string;
  /** @minLength 1 */
  subject: string;
  year: number;
  className: ClassLevel;
  /** @minLength 1 */
  fileUrl: string;
}

export interface GalleryImage {
  id: number;
  imageUrl: string;
  /** @nullable */
  caption: string | null;
  uploadDate: string;
}

export interface GalleryImageInput {
  /** @minLength 1 */
  imageUrl: string;
  caption?: string;
}

export interface Announcement {
  id: number;
  title: string;
  description: string;
  pinned: boolean;
  createdAt: string;
}

export interface AnnouncementInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  pinned?: boolean;
}

export interface AnnouncementUpdate {
  title?: string;
  description?: string;
  pinned?: boolean;
}

export interface ContactMessage {
  id: number;
  fullName: string;
  mobile: string;
  /** @nullable */
  email: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface ContactMessageInput {
  /** @minLength 1 */
  fullName: string;
  /**
     * @minLength 10
     * @maxLength 10
     */
  mobile: string;
  email?: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  message: string;
}

export interface NewsletterSubscriber {
  id: number;
  name: string;
  mobile: string;
  /** @nullable */
  email: string | null;
  createdAt: string;
}

export interface NewsletterInput {
  /** @minLength 1 */
  name: string;
  /**
     * @minLength 10
     * @maxLength 10
     */
  mobile: string;
  email?: string;
}

export interface Notification {
  id: number;
  title: string;
  description: string;
  createdAt: string;
}

export interface PublicStats {
  studentsEnrolled: number;
  expertTeachers: number;
  coursesCount: number;
  successRate: number;
}

export interface AdminDashboardSummary {
  totalStudents: number;
  activeStudents: number;
  terminatedStudents: number;
  totalFeesCollected: number;
  pendingFees: number;
  monthlyCollection: number;
  newAdmissionsThisMonth: number;
}

export type ListNotesParams = {
className?: ClassLevel;
subject?: string;
search?: string;
};

export type ListPapersParams = {
board?: string;
subject?: string;
search?: string;
};

export type ListStudentsParams = {
search?: string;
className?: ClassLevel;
status?: StudentStatus;
sort?: ListStudentsSort;
};

export type ListStudentsSort = typeof ListStudentsSort[keyof typeof ListStudentsSort];


export const ListStudentsSort = {
  newest: 'newest',
  oldest: 'oldest',
  nameAsc: 'nameAsc',
  nameDesc: 'nameDesc',
} as const;