# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contact:
    post:
      operationId: submitContactMessage
      tags: [public]
      summary: Submit a contact enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "400":
          description: Validation error
  /newsletter:
    post:
      operationId: subscribeNewsletter
      tags: [public]
      summary: Subscribe to the newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscriber"
        "400":
          description: Validation error
...
  /notes:
    get:
      operationId: listNotes
      tags: [public]
      summary: List notes (optionally filtered)
      parameters:
        - name: className
          in: query
          schema:
            $ref: "#/components/schemas/ClassLevel"
        - name: subject
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
  /papers:
    get:
      operationId: listPapers
      tags: [public]
      summary: List previous year papers (optionally filtered)
      parameters:
        - name: board
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paper"
...
  /auth/student/register:
    post:
      operationId: registerStudent
      tags: [studentAuth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentRegistrationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Validation error
        "409":
          description: Mobile number already registered
  /auth/student/login:
    post:
      operationId: loginStudent
      tags: [studentAuth]
      summary: Login as a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Invalid credentials
...
  /auth/student/me:
    get:
      operationId: getCurrentStudent
      tags: [studentAuth]
      summary: Get currently logged in student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not logged in
  /student/profile:
    get:
      operationId: getStudentProfile
      tags: [student]
      summary: Get own profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not logged in
    patch:
      operationId: updateStudentProfilePhoto
      tags: [student]
      summary: Update own profile photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfilePhotoUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not logged in
  /student/change-password:
    post:
      operationId: changeStudentPassword
      tags: [student]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "204":
          description: Changed
        "400":
          description: Validation error
        "401":
          description: Not logged in / wrong current password
  /student/fees:
    get:
      operationId: getStudentFees
      tags: [student]
      summary: Get own fee schedule
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fee"
        "401":
          description: Not logged in
  /student/notifications:
    get:
      operationId: getStudentNotifications
      tags: [student]
      summary: Get notifications targeted at this student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: Not logged in
  /auth/admin/login:
    post:
      operationId: loginAdmin
      tags: [adminAuth]
      summary: Login as admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "401":
          description: Invalid credentials
...
  /auth/admin/me:
    get:
      operationId: getCurrentAdmin
      tags: [adminAuth]
      summary: Get currently logged in admin
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "401":
          description: Not logged in
...
components:
  schemas:
...
    ClassLevel:
      type: string
      enum: ["6-8", "9-10", "11-12"]
    StudentStatus:
      type: string
      enum: [active, inactive, suspended, terminated]
    FeeStatus:
      type: string
      enum: [paid, partial, pending]
...
    Student:
      type: object
      required:
        - id
        - studentId
        - admissionNumber
        - fullName
        - fatherName
        - mobile
        - email
        - pinCode
        - schoolName
        - className
        - admissionDate
        - registrationDate
        - status
        - profilePhotoUrl
        - terminationDate
        - terminationReason
        - createdAt
      properties:
        id: { type: integer }
        studentId: { type: string }
        admissionNumber: { type: string }
        fullName: { type: string }
        fatherName: { type: string }
        mobile: { type: string }
        email: { type: ["string", "null"] }
        pinCode: { type: string }
        schoolName: { type: string }
        className:
          $ref: "#/components/schemas/ClassLevel"
        admissionDate: { type: string, format: date }
        registrationDate: { type: string, format: date-time }
        status:
          $ref: "#/components/schemas/StudentStatus"
        profilePhotoUrl: { type: ["string", "null"] }
        terminationDate: { type: ["string", "null"], format: date-time }
        terminationReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    StudentRegistrationInput:
      type: object
      required:
        - fullName
        - fatherName
        - mobile
        - pinCode
        - schoolName
        - className
        - password
      properties:
        fullName: { type: string, minLength: 1 }
        fatherName: { type: string, minLength: 1 }
        mobile: { type: string, minLength: 10, maxLength: 10 }
        email: { type: string }
        pinCode: { type: string, minLength: 6, maxLength: 6 }
        schoolName: { type: string, minLength: 1 }
        className:
          $ref: "#/components/schemas/ClassLevel"
        password: { type: string, minLength: 8 }
    StudentLoginInput:
      type: object
      required: [mobile, password]
      properties:
        mobile: { type: string }
        password: { type: string }
...
    ProfilePhotoUpdate:
      type: object
      required: [profilePhotoUrl]
      properties:
        profilePhotoUrl: { type: string }
    PasswordChangeInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 8 }
...
    Admin:
      type: object
      required: [id, username, mustChangePassword, lastLoginAt]
      properties:
        id: { type: integer }
        username: { type: string }
        mustChangePassword: { type: boolean }
        lastLoginAt: { type: ["string", "null"], format: date-time }
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    Fee:
      type: object
      required:
        - id
        - studentId
        - month
        - year
        - feeAmount
        - discount
        - lateFee
        - amountPaid
        - status
        - paymentDate
        - receiptNumber
        - paymentMode
        - remarks
      properties:
        id: { type: integer }
        studentId: { type: integer }
        month: { type: integer, minimum: 1, maximum: 12 }
        year: { type: integer }
        feeAmount: { type: number }
        discount: { type: number }
        lateFee: { type: number }
        amountPaid: { type: number }
        status:
          $ref: "#/components/schemas/FeeStatus"
        paymentDate: { type: ["string", "null"], format: date }
        receiptNumber: { type: ["string", "null"] }
        paymentMode:
          type: ["string", "null"]
          enum: [cash, upi, bank_transfer, debit_card, credit_card, other, null]
        remarks: { type: ["string", "null"] }
...
    Note:
      type: object
      required: [id, title, subject, className, fileUrl, uploadDate]
      properties:
        id: { type: integer }
        title: { type: string }
        subject: { type: string }
        className:
          $ref: "#/components/schemas/ClassLevel"
        fileUrl: { type: string }
        uploadDate: { type: string, format: date-time }
...
    Paper:
      type: object
      required: [id, title, board, subject, year, className, fileUrl, uploadDate]
      properties:
        id: { type: integer }
        title: { type: string }
        board: { type: string }
        subject: { type: string }
        year: { type: integer }
        className:
          $ref: "#/components/schemas/ClassLevel"
        fileUrl: { type: string }
        uploadDate: { type: string, format: date-time }
...
    ContactMessage:
      type: object
      required: [id, fullName, mobile, email, subject, message, createdAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        mobile: { type: string }
        email: { type: ["string", "null"] }
        subject: { type: string }
        message: { type: string }
        createdAt: { type: string, format: date-time }
    ContactMessageInput:
      type: object
      required: [fullName, mobile, subject, message]
      properties:
        fullName: { type: string, minLength: 1 }
        mobile: { type: string, minLength: 10, maxLength: 10 }
        email: { type: string }
        subject: { type: string, minLength: 1 }
        message: { type: string, minLength: 1 }
    NewsletterSubscriber:
      type: object
      required: [id, name, mobile, email, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        mobile: { type: string }
        email: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    NewsletterInput:
      type: object
      required: [name, mobile]
      properties:
        name: { type: string, minLength: 1 }
        mobile: { type: string, minLength: 10, maxLength: 10 }
        email: { type: string }
    Notification:
      type: object
      required: [id, title, description, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        createdAt: { type: string, format: date-time }
...