# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: number;
  username: string;
  mustChangePassword: boolean;
  /** @nullable */
  lastLoginAt: Date | null;
}