# lib/api-zod/src/generated/types/fee.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeePaymentMode } from './feePaymentMode';
import type { FeeStatus } from './feeStatus';

export interface Fee {
  id: number;
  studentId: number;
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  year: number;
  feeAmount: number;
  discount: number;
  lateFee: number;
  amountPaid: number;
  status: FeeStatus;
  /** @nullable */
  paymentDate: Date | null;
  /** @nullable */
  receiptNumber: string | null;
  /** @nullable */
  paymentMode: FeePaymentMode;
  /** @nullable */
  remarks: string | null;
}