# lib/api-zod/src/generated/types/feeStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type FeeStatus = typeof FeeStatus[keyof typeof FeeStatus];


export const FeeStatus = {
  paid: 'paid',
  partial: 'partial',
  pending: 'pending',
} as const;