# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClassLevel } from './classLevel';
import type { StudentStatus } from './studentStatus';

export interface Student {
  id: number;
  studentId: string;
  admissionNumber: string;
  fullName: string;
  fatherName: string;
  mobile: string;
  /** @nullable */
  email: string | null;
  pinCode: string;
  schoolName: string;
  className: ClassLevel;
  admissionDate: Date;
  registrationDate: Date;
  status: StudentStatus;
  /** @nullable */
  profilePhotoUrl: string | null;
  /** @nullable */
  terminationDate: Date | null;
  /** @nullable */
  terminationReason: string | null;
  createdAt: Date;
}