# lib/api-zod/src/generated/types/studentLoginInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentLoginInput {
  mobile: string;
  password: string;
}